requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
gspread==3.0.1
httplib2==0.11.3
idna==2.7
oauth2client==4.1.2
pipenv==2018.7.1
psycopg2==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pymongo==3.7.1
python-telegram-bot==10.1.0
requests==2.19.1
rsa==3.4.2
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
